+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - New #164: Add `isDirectMessages` field to `Chat` and `ChatFullInfo` types.
 - New #164: Add `parentChat` field to `ChatFullInfo` type.
 - New #164: Add `DirectMessagesTopic` type.
+- New #164: Add `directMessagesTopic` field to `Message` type.
 
 ## 0.9.1 July 4, 2025